增加开发平台路由控制

+++ Module/OpenPlatform/Service/PlatformController.php
@@ -1,11 +1,8 @@
 <?php
 namespace Module\OpenPlatform\Service;
 
-
 use Module\OpenPlatform\Domain\Model\PlatformModel;
-
-
-
+use Zodream\Infrastructure\Http\Request;
 
 class PlatformController extends Controller {
 
@@ -20,14 +17,19 @@ class PlatformController extends Controller {
 
     public function editAction($id) {
         $model = PlatformModel::findOrNew($id);
+        if ($id > 0 && $model->user_id !== auth()->id()) {
+            return $this->redirectWithMessage('./', '应用错误！');
+        }
         return $this->show(compact('model'));
     }
 
-    public function saveAction() {
-        $id = intval(app('request')->get('id'));
-        $data = app('request')->get();
+    public function saveAction(Request $request) {
+        $id = intval($request->get('id'));
+        $data = $request->get();
         unset($data['appid']);
         unset($data['secret']);
+        unset($data['rules']);
+        unset($data['status']);
         if ($id > 0) {
             $model = PlatformModel::where('user_id', auth()->id())
                 ->where('id', $id)->one();
@@ -35,6 +37,7 @@ class PlatformController extends Controller {
             $model = new PlatformModel();
             $model->user_id = auth()->id();
             $model->generateNewId();
+            $model->status = PlatformModel::STATUS_WAITING;
         }
         if (empty($model)) {
             return $this->jsonFailure('应用不存在');
@@ -48,7 +51,7 @@ class PlatformController extends Controller {
     }
 
     public function deleteAction($id) {
-        PlatformModel::where('id', $id)->delete();
+        PlatformModel::where('id', $id)->where('user_id', auth()->id())->delete();
         return $this->jsonSuccess([
             'url' => url('./platform')
         ]);
